Move XDF dialogs into one module
